Moved register_object into init methods Fixed PEP-8 warnings in session
